Move as much items as possible in diagram specific package
