+++ gaphor/diagram/actions/activitynodes.py
