+++ gaphor/diagram/actions/objectnode.py
