+++ gaphor/diagram/actions/tests/test_activitynodes.py
