+++ gaphor/diagram/actions/tests/test_objectnode.py
