+++ gaphor/diagram/components/artifact.py
