+++ gaphor/diagram/components/component.py
