+++ gaphor/diagram/components/connector.py
