+++ gaphor/diagram/components/node.py
