+++ gaphor/diagram/general/comment.py
