+++ gaphor/diagram/general/commentline.py
