+++ gaphor/diagram/general/simpleitem.py
