+++ gaphor/diagram/interactions/interaction.py
