+++ gaphor/diagram/interactions/lifeline.py
