+++ gaphor/diagram/interactions/message.py
