+++ gaphor/diagram/profiles/extension.py
