+++ gaphor/diagram/usecases/actor.py
