+++ gaphor/diagram/usecases/usecase.py
