+++ gaphor/ui/tests/test_handletool.py
@@ -10,9 +10,9 @@ from gi.repository import Gdk, Gtk
 from gaphor import UML
 from gaphor.application import Application
 from gaphor.core import inject
-from gaphor.diagram.actor import ActorItem
-from gaphor.diagram.comment import CommentItem
-from gaphor.diagram.commentline import CommentLineItem
+from gaphor.diagram.usecases.actor import ActorItem
+from gaphor.diagram.general.comment import CommentItem
+from gaphor.diagram.general.commentline import CommentLineItem
 from gaphor.diagram.interfaces import IConnect
 from gaphor.ui.diagramtools import ConnectHandleTool, DiagramItemConnector
 from gaphor.ui.event import DiagramShow
